feat: manage account buttons
