+++ src/main/events/index.ts
@@ -30,6 +30,7 @@ import "./library/remove-game-from-library";
 import "./library/select-game-wine-prefix";
 import "./library/reset-game-achievements";
 import "./misc/open-checkout";
+import "./misc/open-manage-account";
 import "./misc/open-external";
 import "./misc/show-open-dialog";
 import "./misc/get-features";
