+++ src/renderer/src/declaration.d.ts
@@ -29,6 +29,7 @@ import type {
   UserAchievement,
   ComparedAchievements,
   CatalogueSearchPayload,
+  ManageAccountPage,
 } from "@types";
 import type { AxiosProgressEvent } from "axios";
 import type disk from "diskusage";
@@ -187,6 +188,7 @@ declare global {
     /* Misc */
     openExternal: (src: string) => Promise<void>;
     openCheckout: () => Promise<void>;
+    openManageAccount: (page: ManageAccountPage) => Promise<void>;
     getVersion: () => Promise<string>;
     isStaging: () => Promise<boolean>;
     ping: () => string;
