+++ src/renderer/src/pages/settings/settings.tsx
@@ -11,7 +11,7 @@ import {
   SettingsContextConsumer,
   SettingsContextProvider,
 } from "@renderer/context";
-import { SettingsPrivacy } from "./settings-privacy";
+import { SettingsAccount } from "./settings-account";
 import { useUserDetails } from "@renderer/hooks";
 import { useMemo } from "react";
 
@@ -28,7 +28,7 @@ export default function Settings() {
       "Real-Debrid",
     ];
 
-    if (userDetails) return [...categories, t("privacy")];
+    if (userDetails) return [...categories, t("account")];
     return categories;
   }, [userDetails, t]);
 
@@ -53,7 +53,7 @@ export default function Settings() {
               return <SettingsRealDebrid />;
             }
 
-            return <SettingsPrivacy />;
+            return <SettingsAccount />;
           };
 
           return (
